fix: relocate my new method to data_utils
