+++ src/service/python_sdk_service.py
@@ -8,7 +8,7 @@ from typing import Any, Dict, Optional
 from src.config.config_models import SourceConfig
 from src.config.settings import Settings
 from src.service.base_service import BaseSourceService, ServiceError
-from src.utils.api_response import dict_response_key_to_records
+from src.utils.data_utils import dict_response_key_to_records
 from src.utils.dynamic_values import get_resolver, resolve_headers_dict
 from src.utils.logger import get_logger
 from src.utils.redis_context import RedisContextManager
